% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGM.R
\name{BM_fit}
\alias{BM_fit}
\title{Fit the Boltzmann Machine via maximum likelihood}
\usage{
BM_fit(
  V,
  p = ncol(V),
  m = 50,
  N = 1000,
  step_size = 1e-05,
  epsilon = 0.1,
  max_iter = 100
)
}
\arguments{
\item{V}{The input dataset on visible varibales, of dimension nobs x nvars; each row is an observation vector.}

\item{p}{The number of visible variables. Default is ncol(V).}

\item{m}{The number of hidden variables. Defaut value is 50.}

\item{N}{The number of Monte Carlo samples to estimate z(theta). Default is 1000}

\item{step_size}{The step size for the gradient descent. Defaut is 0.00001}

\item{epsilon}{The convergence criterion. Defaut is 0.1}

\item{max_iter}{The maximum number of iterations. Defaut is 100}
}
\value{
Estimate of the parameter theta, a (p+m) x (p+m) matrix.
}
\description{
Fit the Boltzmann Machine via maximum likelihood
}
\examples{
n = 1000
p = 3
rho = 0.5
Sigma = (1-rho)*diag(p) + rho*matrix(1, p, p)
z = rmvnorm(n, sigma = Sigma)
y = matrix(0, n, p)
y[z>0] = 1
step_size = 0.001
k = 1
epsilon = 0.001
N = 1000
max_iter = 1000
res_rbm = RBM_fit(y, p, m = 2*p, N, step_size, epsilon, k, method = "likelihood", max_iter)
res_bm = BM_fit(y, p, m = 2*p, N, step_size = 0.0001, epsilon = 0.03, max_iter)
}
