% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGM.R
\name{BM_reconstruct}
\alias{BM_reconstruct}
\title{Reconstruct the data from the Boltzmann Machine}
\usage{
BM_reconstruct(v, theta, p = ncol(v), m = 50, max_iter = 100, L = 200)
}
\arguments{
\item{v}{The input data on visible variables.}

\item{theta}{The parameter of the Boltzmann Machine.}

\item{p}{The number of visible variables. The default is ncol(v).}

\item{m}{The number of hidden variables. The default is 50.}

\item{max_iter}{The maximum number of iterations. The default is 100.}

\item{L}{The number of samples to be generated. The default is 200.}
}
\value{
The reconstructed data.
}
\description{
Reconstruct the data from the Boltzmann Machine
}
\examples{
p = 3
n = 200
rho = 0.5
Sigma = (1-rho)*diag(p) + rho*matrix(1, p, p)
z = mvtnorm::rmvnorm(n, sigma = Sigma)
y = matrix(0, n, p)
y[z > 0] = 1
m = 2*p
theta_BM = BM_fit(y, p, m = m, N = 100, step_size = 1e-4, epsilon = 0.03, max_iter = 20)
V_recon = BM_reconstruct(y[1, ], theta_BM, p = p, m = m, max_iter = 20, L = 50)
}
