% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGM.R
\name{EGM.CV}
\alias{EGM.CV}
\title{This function function implements the k-fold cross-validation for EGM.fit, produces the optimal lambda and step size, and returns the resulting theta.}
\usage{
EGM.CV(
  X,
  model,
  theta = 0,
  lambda = 0,
  NMC = 500,
  step_size = 0.003,
  epsilon = 1e-04,
  max_iter = 3000,
  k = 10,
  par = FALSE,
  max_cores = 4
)
}
\arguments{
\item{X}{Input dataset, of dimension n x p; each row is an observation vector.}

\item{model}{The model to be fitted. The model must be either "pgm" or "ising".}

\item{theta}{The initial value of the parameter. The default is 0.}

\item{lambda}{A gird of lambda (penalty parameter) values used for the EGM.fit.}

\item{NMC}{The number of Monte Carlo samples to approximate the gradient of the log-likelihood. The default is 500.}

\item{step_size}{A grid of step size values for the gradient descent. The default is 0.003.}

\item{epsilon}{The convergence criterion. Default is 0.0001.}

\item{max_iter}{The maximum number of iterations. The default is 3000.}

\item{k}{The number of folds for cross-validation. The default is 10.}

\item{par}{Logical. If TRUE, the function will run in parallel. The default is FALSE.}

\item{max_cores}{The maximum number of cores to be used for parallel computation. The default is 4.}
}
\value{
A list containing the following elements: 
\itemize{
  \item theta_lasso: The estimate of parameter theta, a p x p matrix.
  \item lambda: The optimal lambda value.
  \item step_size: The optimal step size value.
}#'
}
\description{
This function function implements the k-fold cross-validation for EGM.fit, produces the optimal lambda and step size, and returns the resulting theta.
}
\examples{

theta0 = gen_theta0(p = 10, omega = 0.8, eta = 0.05, coupling = "negative")
X = IsingSim(n = 100, theta = theta0, max_iter = 1000)
res = EGM.CV(X = X, model = "ising", lambda = c(1:15), step_size = c(0.001, 0.0005, 0.0001), k = 10, par = TRUE, max_cores = 50)
}
