% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGM.R
\name{EGM.fit}
\alias{EGM.fit}
\title{Fit the likelihood-based Exponential Graphical Model via maximum likelihood or penalized maximum likelihood.}
\usage{
EGM.fit(
  X,
  model,
  theta = 0,
  lambda = NULL,
  penalty = "none",
  NMC = 500,
  step_size = 0.003,
  epsilon = 1e-05,
  max_iter = 3000
)
}
\arguments{
\item{X}{Input dataset of dimension n x p; each row is an observation vector.}

\item{model}{The model to be fitted. The model must be either "pgm" or "ising".}

\item{theta}{The initial value of the parameter. The default is 0.}

\item{lambda}{The penalty parameter. The default is NULL.}

\item{penalty}{The penalty type. The penalty must be either "none" or "l1". The default is "none".}

\item{NMC}{The number of Monte Carlo samples for approximate the gradient of the log-likelihood. The default is 500.}

\item{step_size}{The step size for the gradient descent. The default is 0.003.}

\item{epsilon}{The convergence criterion. The default is 0.00001.}

\item{max_iter}{The maximum number of iterations. The default is 3000.}
}
\value{
Estimate of parameter theta, a p x p matrix.
}
\description{
Fit the likelihood-based Exponential Graphical Model via maximum likelihood or penalized maximum likelihood.
}
\examples{
p = 3
n = 100
theta0 = matrix(-0.8, p, p)
X_3 = IsingSim(n, theta0, 1000)
EGM.fit(X = X_3, model = "ising")
}
