% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGM.R
\name{RBM_fit}
\alias{RBM_fit}
\title{Fit the Restricted Boltzmann Machine via maximum likelihood method.}
\usage{
RBM_fit(
  V,
  p = ncol(V),
  m = 30,
  N = 1000,
  step_size = 1e-05,
  epsilon = 0.06,
  k = 1,
  method = "likelihood",
  max_iter = 1000
)
}
\arguments{
\item{V}{The input dataset on visible varibales, of dimension n x p; each row is an observation vector.}

\item{p}{The number of visible variables.}

\item{m}{The number of hidden variables.}

\item{N}{The number of Monte Carlo samples to estimate z(theta).}

\item{step_size}{The step size for the gradient descent.}

\item{epsilon}{The convergence criterion.}

\item{k}{The number of Gibbs samples in CD.}

\item{method}{The method to be used for fitting the RBM. The method must be either "likelihood" or "CD". Default is "likelihood".}

\item{max_iter}{The maximum number of iterations.}
}
\value{
Estimate of the parameter theta, a (p+m) x (p+m) matrix.
}
\description{
Fit the Restricted Boltzmann Machine via maximum likelihood method.
}
\examples{
n = 1000
p = 3
rho = 0.5
Sigma = (1-rho)*diag(p) + rho*matrix(1, p, p)
z = mvtnorm::rmvnorm(n, sigma = Sigma)
y = matrix(0, n, p)
y[z>0] = 1
step_size = 0.001
k = 1
epsilon = 0.001
N = 1000
max_iter = 1000
res_rbm = RBM_fit(y, p, m = 2*p, N, step_size, epsilon, k, method = "likelihood", max_iter)
res_bm = BM_fit(y, p, m = 2*p, N, step_size = 0.0001, epsilon = 0.03, max_iter)
}
