% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGM.R
\name{RBM_reconstruct}
\alias{RBM_reconstruct}
\title{Reconstruct the data from the Restricted Boltzmann Machine}
\usage{
RBM_reconstruct(v, theta, p = ncol(v), m = 30)
}
\arguments{
\item{v}{The input data on visible variables.}

\item{theta}{The parameter of the Restricted Boltzmann Machine.}

\item{p}{The number of visible variables.}

\item{m}{The number of hidden variables.}
}
\value{
The reconstructed data.
}
\description{
Reconstruct the data from the Restricted Boltzmann Machine
}
