% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGM.R
\name{BM_reconstruct}
\alias{BM_reconstruct}
\title{Reconstruct the data from the Boltzmann Machine}
\usage{
BM_reconstruct(v, theta, p = ncol(v), m = 50, max_iter = 100, L = 200)
}
\arguments{
\item{v}{The input data on visible variables.}

\item{theta}{The parameter of the Boltzmann Machine.}

\item{p}{The number of visible variables. The default is ncol(v).}

\item{m}{The number of hidden variables. The default is 50.}

\item{max_iter}{The maximum number of iterations. The default is 100.}

\item{L}{The number of samples to be generated. The default is 200.}
}
\value{
The reconstructed data.
}
\description{
Reconstruct the data from the Boltzmann Machine
}
\examples{
V_recon = BM_reconstruct(v, theta_BM, p, m, max_iter, L)
}
