\name{ExponentialGM-package}
\alias{ExponentialGM-package}
\alias{ExponentialGM}
\docType{package}
\title{
\packageTitle{ExponentialGM}
}
\description{
\packageDescription{ExponentialGM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ExponentialGM}
\packageIndices{ExponentialGM}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ExponentialGM}

Maintainer: \packageMaintainer{ExponentialGM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
